Update Main layout for responsive detail view
- Rename existing Main to Main.old
- Rename NewMain to Main
- Update pages to use updated Main component
- Update single col pags to use MainSingle
