+++ ui/src/components/Main.tsx
@@ -1,53 +1,54 @@
-import * as React from "react";
 import styled, { css } from "styled-components";
 
-import { up, down } from "styled-breakpoints";
 import { SCREENS } from "../media";
 
-export const Col = styled.div`
-  width: 100%;
-
-  ${SCREENS.Up.Desktop} {
-    max-width: 29.75rem;
-    & + & {
-      margin-left: 1.375rem;
-    }
-  }
-`;
-
-type DetailColProps = {
-  active?: boolean;
-  onClick?(): void;
-};
-
-export const DetailCol = styled(Col)<DetailColProps>`
-  ${SCREENS.Down.Tablet} {
-    position: fixed;
-    top: 0;
-    left: 0;
-    bottom: 0;
-    right: 0;
-    padding: 20px;
-    background: rgba(0, 0, 0, 0.2);
-    display: ${(props) => (props.active ? "block" : "none")};
-  }
-`;
-
-export const StyledMain = styled.div`
-  grid-area: main;
-  padding: 0 1rem;
-
-  ${SCREENS.Up.Desktop} {
-    display: flex;
-    justify-content: center;
-  }
-`;
-
 type MainProps = {
-  children: React.ReactNode;
+  active?: boolean;
 };
 
-export const Main: React.FC<MainProps> = function (props) {
-  const { children } = props;
-  return <StyledMain>{children}</StyledMain>;
-};
+export const Main = styled.div<MainProps>`
+  outline: 1px solid red;
+  outline-offset: -1px;
+  grid-area: main;
+  display: grid;
+  grid-template-columns: 100vw 100vw;
+  grid-template-rows: 2.875rem 2.875rem auto;
+  grid-row-gap: 1rem;
+  transition: transform 200ms ease-in-out;
+  position: relative;
+
+  grid-template-areas:
+    "header-left detail"
+    "header-right detail"
+    "list detail";
+
+  ${SCREENS.Down.Tablet} {
+    ${(props) =>
+      props.active &&
+      css`
+        transform: translateX(-100vw);
+      `}
+  }
+
+  ${SCREENS.Up.Desktop} {
+    padding: 0 1rem;
+    justify-content: center;
+    grid-template-rows: 2.875rem auto;
+    grid-template-columns: repeat(2, 29.75rem);
+    grid-column-gap: 1.375rem;
+    grid-row-gap: 1.375rem;
+    grid-template-areas:
+      "header-left header-right"
+      "list detail";
+  }
+`;
+
+export const MainSingle = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  grid-area: main;
+  max-width: 30rem;
+  width: 100%;
+  margin: 0 auto;
+`;
