+++ ui/src/screens/Home/HomeScreen.tsx
@@ -2,46 +2,42 @@ import * as React from "react";
 import { Link } from "react-router-dom";
 
 import Button from "../../components/Button";
-import { Main, Col } from "../../components/Main";
+import { MainSingle } from "../../components/Main";
 
 const HomeScreen: React.FC = function () {
   return (
-    <Main>
-      <Col>
+    <MainSingle>
+      <p>
+        <Link to="/jobs">
+          <Button primary>Go to jobs board</Button>
+        </Link>
+      </p>
+      <p>
+        <Link to="/talents">
+          <Button primary>Go to talents board</Button>
+        </Link>
+      </p>
+      <p>
+        <Link to="/favourites">
+          <Button secondary>Favourites</Button>
+        </Link>
+      </p>
+      <div>
+        <h2>Buttons</h2>
         <p>
-          <Link to="/jobs">
-            <Button primary>Go to jobs board</Button>
-          </Link>
+          <Button>Default Button</Button>
         </p>
         <p>
-          <Link to="/talents">
-            <Button primary>Go to talents board</Button>
-          </Link>
+          <Button primary>Primary Button</Button>
         </p>
         <p>
-          <Link to="/favourites">
-            <Button secondary>Favourites</Button>
-          </Link>
+          <Button secondary>Secondary Button</Button>
         </p>
-      </Col>
-      <Col>
-        <div>
-          <h2>Buttons</h2>
-          <p>
-            <Button>Default Button</Button>
-          </p>
-          <p>
-            <Button primary>Primary Button</Button>
-          </p>
-          <p>
-            <Button secondary>Secondary Button</Button>
-          </p>
-          <p>
-            <Button active>Active Button</Button>
-          </p>
-        </div>
-      </Col>
-    </Main>
+        <p>
+          <Button active>Active Button</Button>
+        </p>
+      </div>
+    </MainSingle>
   );
 };
