+++ ui/src/screens/Jobs/JobsScreen.tsx
@@ -1,60 +1,43 @@
 import * as React from "react";
-import styled from "styled-components";
-import { Link } from "react-router-dom";
 
-import { Main, Col, DetailCol } from "../../components/Main";
+import { Main } from "../../components/Main";
+import Search from "../../shared/components/Search";
 import CategorySelector from "../../shared/components/CategorySelector";
 import Post from "../../shared/components/Post";
 import PostDetail from "../../shared/components/PostDetail";
 import PostDetailPlaceholder from "../../shared/components/PostDetailPlaceholder";
 import PostsContainer from "../../shared/components/PostsContainer";
-import { SCREENS } from "../../media";
+import DetailsContainer from "../../shared/components/DetailsContainer";
 
-import { CategoryTypes, PostInterface } from "../../types";
+import { PostInterface } from "../../types";
 
 import usePostsReducer from "./hooks/useJobsReducer";
-
-const DetailsContainer = styled.div`
-  flex: 1;
-
-  ${SCREENS.Up.Desktop} {
-    .back {
-      display: none;
-    }
-  }
-`;
 
 const JobsScreen: React.FC = function () {
   const [state, actions] = usePostsReducer();
   const { toggleFavouriteJob } = actions;
 
   return (
-    <Main>
-      <Col>
-        <PostsContainer>
-          {state.jobs.map((post: PostInterface) => (
-            <Post
-              category="jobs"
-              active={post.active}
-              key={post.id}
-              data={post}
-              handleFavouriteToggle={(e): void => {
-                e.stopPropagation();
-                toggleFavouriteJob(post);
-              }}
-            />
-          ))}
-        </PostsContainer>
-      </Col>
-      <DetailCol active={Boolean(state.activeJob)}>
-        <CategorySelector category="jobs" />
-        <DetailsContainer>
-          <Link className="back" to="/jobs">
-            <button>Back</button>
-          </Link>
-          {state.activeJob ? <PostDetail data={state.activeJob} /> : <PostDetailPlaceholder />}
-        </DetailsContainer>
-      </DetailCol>
+    <Main active={Boolean(state.activeJob)}>
+      <Search />
+      <CategorySelector category="jobs" />
+      <PostsContainer>
+        {state.jobs.map((post: PostInterface) => (
+          <Post
+            category="jobs"
+            active={post.active}
+            key={post.id}
+            data={post}
+            handleFavouriteToggle={(e): void => {
+              e.stopPropagation();
+              toggleFavouriteJob(post);
+            }}
+          />
+        ))}
+      </PostsContainer>
+      <DetailsContainer active={Boolean(state.activeJob)}>
+        {state.activeJob ? <PostDetail data={state.activeJob} /> : <PostDetailPlaceholder />}
+      </DetailsContainer>
     </Main>
   );
 };
