+++ ui/src/shared/components/CategorySelector.tsx
@@ -3,16 +3,21 @@ import { Link } from "react-router-dom";
 import styled from "styled-components";
 
 import Button from "../../components/Button";
+import { SCREENS } from "../../media";
 import { CategoryTypes } from "../../types";
 
 const Container = styled.div`
   display: flex;
+  grid-area: header-right;
   margin-bottom: 1rem;
   * + * {
     margin-left: 0.5rem;
   }
   ${Button} {
     width: 11.5rem;
+  }
+  ${SCREENS.Down.Tablet} {
+    padding: 0 1rem;
   }
 `;
